CORE-16919: Move tracing implementation into :libs:tracing-impl module. (#4590)
