+++ applications/workers/release/db-worker/build.gradle
@@ -33,12 +33,14 @@ dependencies {
     implementation 'net.corda:corda-config-schema'
 
     runtimeOnly "com.typesafe:config:$typeSafeConfigVersion"
+    runtimeOnly "org.apache.aries.spifly:org.apache.aries.spifly.dynamic.framework.extension:$ariesDynamicFrameworkExtensionVersion"
     runtimeOnly "org.apache.felix:org.apache.felix.scr:$felixScrVersion"
     runtimeOnly "org.osgi:org.osgi.service.component:$osgiServiceComponentVersion"
     runtimeOnly "org.osgi:org.osgi.util.function:$osgiUtilFunctionVersion"
     runtimeOnly "org.osgi:org.osgi.util.promise:$osgiUtilPromiseVersion"
 
     runtimeOnly project(':libs:messaging:kafka-message-bus-impl')
+    runtimeOnly project(':libs:tracing-impl')
 
     testImplementation 'org.osgi:osgi.core'
     testImplementation "org.mockito:mockito-core:$mockitoVersion"
