+++ applications/workers/release/p2p-link-manager-worker/build.gradle
@@ -28,8 +28,10 @@ dependencies {
     implementation "info.picocli:picocli:$picocliVersion"
 
     runtimeOnly project(":libs:messaging:kafka-message-bus-impl")
+    runtimeOnly project(':libs:tracing-impl')
 
     runtimeOnly "org.apache.logging.log4j:log4j-slf4j-impl:$log4jVersion"
+    runtimeOnly "org.apache.aries.spifly:org.apache.aries.spifly.dynamic.framework.extension:$ariesDynamicFrameworkExtensionVersion"
     runtimeOnly "org.apache.felix:org.apache.felix.scr:$felixScrVersion"
     runtimeOnly "org.osgi:org.osgi.service.component:$osgiServiceComponentVersion"
     runtimeOnly "org.osgi:org.osgi.util.function:$osgiUtilFunctionVersion"
