+++ libs/tracing-impl/build.gradle
@@ -1,0 +1,25 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.common-library'
+}
+description 'Tracing Implementation'
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+    compileOnly "biz.aQute.bnd:biz.aQute.bnd.annotation:$bndlibVersion"
+
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
+    implementation project(':libs:tracing')
+
+    implementation("io.zipkin.reporter2:zipkin-sender-urlconnection:$zipkinVersion")
+    implementation("io.zipkin.brave:brave-context-slf4j:$braveVersion")
+    implementation("io.zipkin.brave:brave-instrumentation-servlet:$braveVersion")
+
+    implementation "io.javalin:javalin-osgi:$javalinVersion"
+    constraints {
+        implementation("org.eclipse.jetty:jetty-server:$jettyVersion") {
+            because 'Javalin uses an older version of Jetty which is exposed to CVE-2023-26048 and CVE-2023-26049. ' +
+                    'This might be resolved in the future versions of Javalin.'
+        }
+    }
+}
