+++ libs/tracing-impl/src/main/kotlin/net/corda/tracing/brave/BraveTraceContext.kt
@@ -1,4 +1,4 @@
-package net.corda.tracing.impl
+package net.corda.tracing.brave
 
 import brave.Span
 import brave.Tracer
