+++ libs/tracing/src/main/kotlin/net/corda/tracing/TracingServiceFactory.kt
@@ -1,0 +1,5 @@
+package net.corda.tracing
+
+interface TracingServiceFactory {
+    fun create(serviceName: String, zipkinHost: String?, samplesPerSecond: String?): TracingService
+}
