+++ settings.gradle
@@ -354,6 +354,7 @@ include 'libs:serialization:serialization-kryo:cpks:serializable-cpk-two'
 include 'libs:uniqueness:common'
 include 'libs:utilities'
 include 'libs:tracing'
+include 'libs:tracing-impl'
 include 'libs:virtual-node:virtual-node-common'
 include 'libs:virtual-node:cpi-datamodel'
 include 'libs:virtual-node:cpi-upload-endpoints'
